// src/app/models/horario.model.ts
export class Horario {
    siempreAbierto: boolean;
    Lunes?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
    Martes?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
    Miercoles?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
    Jueves?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
    Viernes?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
    Sabado?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
    Domingo?: {
        abierto: boolean;
        horaInicio?: string;
        horaSalida?: string;
    };
}